Fixed lighting bug
The bug was caused by not clearing the texture that the lights were
being drawn to. Repeated draws of translucent pixels caused weird black
circles to form around lights. This is now fixed.
Also, now all lights are static and only get redrawn when the light
editor is open and a mouse button is down.